Move Int24 to util package
