+++ core/src/main/scala/me/maciejb/snappyflows/util/Int24.scala
@@ -1,4 +1,4 @@
-package me.maciejb.snappyflows
+package me.maciejb.snappyflows.util
 
 import akka.stream.io.ByteStringParser.ByteReader
 import akka.util.ByteString
